Add ASCII decoded versions of OP-2 compiler
This was derived from the compiler import from the previous commit. This
doesn't use DecodeASCII or something within Oberon, which probably would
be more accurate. Instead, I performed this conversion with command line
utilities:
  for i in OP*.Mod; do
    tr -cd '\11\12\15\40-\176' < $i | tr '\r' '\n' > ../FullInstall_ASCII/$i
  done

Which removed all the non-printing characters and tranlated carriage return
to linefeed. The end result required a little bit of hand editing. The
resulting code is readable now, although there may be formatting
instructions encoded in Oberon text docs that I missed in a trivial
pass.